SBERDOMA-1028 moved delete from "update" page to property page
Now it is consistent with other pages.
